// src/types.ts
export interface UseRiveOptions {
  useDevicePixelRatio: boolean;
  customDevicePixelRatio: number;
  fitCanvasToArtboardHeight: boolean;
  useOffscreenRenderer: boolean;
  shouldResizeCanvasToContainer: boolean;
  shouldUseIntersectionObserver?: boolean;
}
// ...
export interface Dimensions {
  width: number;
  height: number;
}